refactor: rename useHot import and move implementation to useHotkey.ts
